pyproject.toml
[tool.poetry]
name = "receiver"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pydantic = "^1.10.7"
black = "^23.3.0"
numpy = "^1.24.2"
pyserial = "^3.5"
pyside6 = "^6.5.0"
plotly = "^5.14.1"
pandas = "^2.0.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
